Remove `-release` from app name for digital marketplace apps
During our deployment of an application called app, we have a new
application spun up called `app-release`. When this app is happy
we take down the already running `app` and then rename our new app
to replace it.

The new application will begin streaming metrics under the application
name `app-release` but we don't want these shown in our
graphana dashboards so we remove the prefix. Even though the app
is renamed, the metrics appear to stream under the two names
(`app` and `app-release`), alternating after each deployment. It is
not entirely clear why.

This fix is intended to be a short run solution before we talk to
the PaaS team about our issue and come up with a longer term
solution.